Backend/requirements.txt
aiosqlite==0.21.0
alembic==1.15.2
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
greenlet==3.2.2
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing_extensions==4.13.2
uvicorn==0.34.2
websockets==15.0.1
